requirements.txt
openai>=1.30.0
pdfplumber>=0.11.0
python-docx>=1.1.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.2.0
python-dotenv>=1.0.0
rich>=13.7.0
pydantic>=2.7.0
feedparser>=6.0.11
